Change the test framework
The testing modules only export one API that should be called by the
main module, the register.

In the register API, the testing modules' start and show function is
given by register these function into the global list.

The main module will call the start and show and loop through the
global list to call all the registered modules.